refactor: create route module
